Refactor module for changes in remark, retext
Been a while! remark now includes `remark-range` internally, so
there’s no reason to depend on it.  Additionally, changes in unified
resulted in the tree no longer being patched on `file`s.

This update ensures everything is nice and clean and updated again.
The interface also changed: the first parameter is now a node, before
the file.

Related to wooorm/remark-retext#1.